Simple Title Exploration
Updated to create training and dev sets (not randomly mixed for now),
and did simple analysis on the posting titles a la Assignment 2 to
start getting a feel for the analysis. My ROC-AUC on just the dev data
was worse than our current score, but that’s probably not surprising.